+++ src/Game/Shared/Components/Components.js
@@ -1,11 +1,13 @@
 /** @format */
 
-const commonJsData = require('./Data');
+const commonJsPack = require('./Pack');
 const commonJsTHREEUtils = require('./THREEUtils');
+const commonJsConstants = require('./Constants');
 const commonJsRoutine = require('./Routine');
 
 module.exports = {
-  Data: commonJsData,
+  Pack: commonJsPack,
   THREEUtils: commonJsTHREEUtils,
   Routine: commonJsRoutine,
+  Constants: commonJsConstants,
 };
