+++ src/Game/Shared/Components/Constants.js
@@ -1,4 +1,5 @@
 /** @format */
+
 module.exports = Object.freeze({
   WEBSOCKET: {
     MSG_TYPES: {
@@ -15,25 +16,5 @@ module.exports = Object.freeze({
       QUERY_AVATAR_GO: 'query_avatar_go',
       SAVE_AVATAR_GO: 'save_avatar_go',
     },
-  },
-
-  //THREAD TODO move this to worldthread and rename this file Constants
-
-  pack(obj) {
-    let OString = JSON.stringify(obj);
-    let SABuffer = new SharedArrayBuffer(
-      Int32Array.BYTES_PER_ELEMENT * OString.length
-    );
-    let sArray = new Int32Array(SABuffer);
-
-    for (let i = 0; i < OString.length; i++) {
-      sArray[i] = OString.charCodeAt(i);
-    }
-
-    return sArray;
-  },
-  unpack(array) {
-    let str = String.fromCharCode.apply(this, array);
-    return JSON.parse(str);
   },
 });
